requirements.txt
flask
boto3
face_recognition
numpy
opencv-python-headless
python-dotenv